form data and cart package

+++ backend/database/migrations/2022_03_08_113656_create_products_prices_table.php
@@ -18,7 +18,7 @@ class CreateProductsPricesTable extends Migration
             $table->integer('product_id');
             $table->integer('storage_id');
             $table->integer('condition_id');
-            $table->float('price')->nullable();
+            $table->integer('price')->nullable();
         });
     }
